@@ -10,11 +10,11 @@ class MicrophoneMeter extends StatefulWidget {
 
 class _MicrophoneMeterState extends State<MicrophoneMeter> {
   final _pd = FlutterPd.instance;
-  PdFileHandle _pdFileHandle;
+  late PdFileHandle _pdFileHandle;
 
   final _assetPath = 'assets/microphone_level.pd';
 
-  Stream<FloatEvent> _pdEvent;
+  late Stream<FloatEvent> _pdEvent;
 
   @override
   void initState() {
@@ -24,7 +24,7 @@ class _MicrophoneMeterState extends State<MicrophoneMeter> {
 
   @override
   void dispose() {
-    _pdFileHandle?.close();
+    _pdFileHandle.close();
     _pd.stopPd();
     super.dispose();
   }
@@ -59,7 +59,7 @@ class _MicrophoneMeterState extends State<MicrophoneMeter> {
           builder: (context, snapshot) {
             if (snapshot.hasData) {
               final event = snapshot.data;
-              return Text('Microphone level: ${event.value}db');
+              return Text('Microphone level: ${event?.value}db');
             } else if (snapshot.hasError) {
               return Text('Error: ${snapshot.error}');
             } else {
