@@ -1,4 +1,4 @@
-package com.niusounds.flutter_pd_example
+package com.andresbrocco.flutter_pd_example
 
 import io.flutter.embedding.android.FlutterActivity
 
